correccion 0.2

+++ src/app/dashboard/pages/users/users.component.ts
@@ -18,52 +18,26 @@ export class UsersComponent {
 
   constructor(
     private matDialog: MatDialog,
-    private usersService: UsersService // MockUsersService // private usersBetterService: UsersBetterService
-  ) {
-    this.users$ = this.usersService.getUsers();
-    // console.log(this.usersBetterService.getUsersWithProduct());
-  }
-
-  addUser(): void {
-    this.matDialog
-      .open(UsersDialogComponent)
+    private usersService: UsersService 
+  ) { this.users$ = this.usersService.getUsers();}
+  addUser(): void { this.matDialog 
+    .open(UsersDialogComponent)
       .afterClosed()
       .subscribe({
         next: (v) => {
-          if (!!v) {
-            this.users$ = this.usersService.createUser(v);
-          }
-        },
-      });
+          if (!!v) {this.users$ = this.usersService.createUser(v);}
+        }
+      })
   }
 
   onEditUser(user: User): void {
     this.matDialog
       .open(UsersDialogComponent, {
         data: user,
+      }).afterClosed().subscribe({next: (v) => {if (!!v) { this.users$ = this.usersService.updateUser(user.id, v)}
+        }
       })
-      .afterClosed()
-      .subscribe({
-        next: (v) => {
-          if (!!v) {
-            this.users$ = this.usersService.updateUser(user.id, v);
-            // CREANDO UNA COPIA DEL ARRAY ACTUAL
-            // const arrayNuevo = [...this.users];
-            // const indiceToEdit = arrayNuevo.findIndex((u) => u.id === user.id);
-            // arrayNuevo[indiceToEdit] = { ...arrayNuevo[indiceToEdit], ...v };
-            // this.users = [...arrayNuevo];
-            // this.users = this.users.map((u) =>
-            //   u.id === user.id ? { ...u, ...v } : u
-            // );
-          }
-        },
-      });
   }
-
-  onDeleteUser(userId: number): void {
-    if (confirm('Esta seguro?')) {
-      this.users$ = this.usersService.deleteUser(userId);
-      // this.users = this.users.filter((u) => u.id !== userId);
-    }
+  onDeleteUser(userId: number): void { if (confirm('Esta seguro?')){this.users$ = this.usersService.deleteUser(userId)}
   }
 }
